signal scope performance upgrade
keep track of the plotted time interval in SignalData so that the size()
and value() method return a subview of the data instead of the entire
history.  Qwt is supposed to clip data points that fall outside the plot
window, but it doesn't seem to be working, so now SignalData does the
clipping itself.  This is a bit hackish, but it performs well.